test(core-components): consolidate prompt template regression spec
Move 5 prompt-template tests from ui-ux/ to core-components/, add a new
modal-persistence test, and surface coverage in QA-CHECKLIST §3.2.
